+++ src/Kiota.Builder/LanguageInformation.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Text.Json;
 using System.Text.Json.Nodes;
 using System.Text.Json.Serialization;
+using Kiota.Builder.Configuration;
 using Kiota.Builder.Extensions;
 using Microsoft.OpenApi;
 
@@ -40,7 +41,6 @@ public record LanguageInformation : IOpenApiSerializable
         writer.WriteStartObject();
         writer.WriteProperty(nameof(MaturityLevel).ToFirstCharacterLowerCase(), MaturityLevel.ToString());
         writer.WriteProperty(nameof(SupportExperience).ToFirstCharacterLowerCase(), SupportExperience.ToString());
-        writer.WriteProperty(nameof(DependencyInstallCommand).ToFirstCharacterLowerCase(), DependencyInstallCommand);
         writer.WriteOptionalCollection(nameof(Dependencies).ToFirstCharacterLowerCase(), Dependencies, callback);
         writer.WriteProperty(nameof(ClientClassName).ToFirstCharacterLowerCase(), ClientClassName);
         writer.WriteProperty(nameof(ClientNamespaceName).ToFirstCharacterLowerCase(), ClientNamespaceName);
@@ -59,18 +59,14 @@ public record LanguageInformation : IOpenApiSerializable
                 if (entry is not null)
                     extension.Dependencies.Add(LanguageDependency.Parse(entry));
         }
-        if (rawObject.TryGetPropertyValue(nameof(DependencyInstallCommand).ToFirstCharacterLowerCase(), out var installCommand) && installCommand is JsonValue stringValue)
-        {
-            extension.DependencyInstallCommand = stringValue.GetValue<string>();
-        }
         // not parsing the maturity level on purpose, we don't want APIs to be able to change that
         if (rawObject.TryGetPropertyValue(nameof(ClientClassName).ToFirstCharacterLowerCase(), out var clientClassName) && clientClassName is JsonValue clientClassNameValue)
         {
-            extension.ClientClassName = clientClassNameValue.GetValue<string>();
+            extension.ClientClassName = GenerationConfiguration.SanitizeClientClassName(clientClassNameValue.GetValue<string>(), string.Empty);
         }
         if (rawObject.TryGetPropertyValue(nameof(ClientNamespaceName).ToFirstCharacterLowerCase(), out var clientNamespaceName) && clientNamespaceName is JsonValue clientNamespaceNameValue)
         {
-            extension.ClientNamespaceName = clientNamespaceNameValue.GetValue<string>();
+            extension.ClientNamespaceName = GenerationConfiguration.SanitizeClientNamespaceName(clientNamespaceNameValue.GetValue<string>(), string.Empty);
         }
         if (rawObject.TryGetPropertyValue(nameof(StructuredMimeTypes).ToFirstCharacterLowerCase(), out var structuredMimeTypes) && structuredMimeTypes is JsonArray structuredMimeTypesValue)
         {